requirements.txt
alembic==1.13.1
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
blinker==1.8.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
crispy==0.8.0
crispy-bootstrap4==2024.10
cycler==0.12.1
Django==5.1.2
django-allauth==65.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-nested-routers==0.94.1
drf-yasg==1.21.10
fabio==2024.9.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
git-filter-repo==2.47.0
greenlet==3.0.3
gunicorn==23.0.0
h5py==3.13.0
hdf5plugin==5.0.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.8
kombu==5.5.3
lxml==5.3.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
redis==6.0.0
requests==2.32.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
silx==2.2.1
six==1.17.0
SQLAlchemy==2.0.29
sqlparse==0.5.1
typing_extensions==4.11.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
xraydb==4.5.6
gunicorn
